requirements.txt
kafka-python
flask
flask_cors